refactor(adapter): extract path matching from the git adapter into its own module
