+++ src/adapter/GitAdapter.ts
@@ -37,18 +37,10 @@ import {
   type ObjectId,
   openRepository,
   type Repository,
-  toSimilarityPercent,
 } from '@scolladon/tsgit'
 
-import { TAB } from '../constant/cliConstants.js'
 import { UTF8_ENCODING } from '../constant/fsConstants.js'
-import {
-  ADDITION,
-  DELETION,
-  HEAD,
-  MODIFICATION,
-  RENAMED,
-} from '../constant/gitConstants.js'
+import { HEAD } from '../constant/gitConstants.js'
 import type { Config } from '../types/config.js'
 import type { FileGitRef } from '../types/git.js'
 import { pushAll } from '../utils/arrayUtils.js'
@@ -63,10 +55,18 @@ import {
   type GitBlobReader,
   SIZE_THRESHOLD,
 } from './gitBlobReader.js'
+import {
+  buildLiteralMatcher,
+  buildPathspecMatcher,
+  hasRootScope,
+  inScope,
+  nonRootScopes,
+  ROOT_PATHS,
+  toDiffLines,
+} from './pathMatching.js'
 import { TreeIndex } from './treeIndex.js'
 import { mapTsgitError } from './tsgitErrorMap.js'
 
-const ROOT_PATHS = new Set(['', '.', './'])
 // walkTree yields directories and gitlinks too; only blob-bearing modes
 // belong in the path -> blob id index (ls-tree -r parity).
 const BLOB_MODES = new Set(['100644', '100755', '120000'])
@@ -82,8 +82,6 @@ const BLOB_MODES = new Set(['100644', '100755', '120000'])
 // every call and nothing here ever reads or mutates shared state for
 // either.
 export type DiffScopeVerdict = { changesSeen: number; linesYielded: number }
-
-const GITLINK_MODE = '160000'
 
 // Maps sgd's whitespace-ignoring diff request onto tsgit's data-mode
 // whitespace knobs: whitespace-only modifications drop out of the TreeDiff.
@@ -608,110 +606,4 @@ const accumulatePointer = async (
     assertWithinPointerCap(length)
   }
   return Buffer.concat(parts, length)
-}
-
-const inScope = (path: string, scopes: readonly string[]): boolean =>
-  scopes.some(
-    scope =>
-      ROOT_PATHS.has(scope) || path === scope || path.startsWith(`${scope}/`)
-  )
-
-const keepSide = (
-  mode: string,
-  path: string,
-  scopes: readonly string[]
-): boolean =>
-  mode !== GITLINK_MODE && (scopes.length === 0 || inScope(path, scopes))
-
-const hasRootScope = (scopes: readonly Pathspec[]): boolean =>
-  scopes.some(scope => ROOT_PATHS.has(scope))
-
-const nonRootScopes = (scopes: readonly Pathspec[]): Pathspec[] =>
-  scopes.filter(scope => !ROOT_PATHS.has(scope))
-
-// git prints rename similarity as a zero-padded three-digit percent (R087).
-const similarityPercent = (similarity: { score: number }): string =>
-  String(toSimilarityPercent(similarity.score)).padStart(3, '0')
-
-// The facade diff takes no pathspec, so `-- <source>` scoping is replicated
-// per side. Gitlink changes are skipped (submodule pointer moves are not
-// deployable metadata) and `type-change`/`copy` entries are dropped for
-// parity with the subprocess `--diff-filter=AMD(R)`. A rename with only one
-// side in scope degrades to that side's A/D line, matching what the
-// subprocess pathspec does to a broken rename pair.
-function* toDiffLines(
-  change: DiffChange,
-  scopes: readonly string[]
-): Generator<string> {
-  switch (change.type) {
-    case 'add':
-      if (keepSide(change.newMode, change.newPath, scopes)) {
-        yield `${ADDITION}${TAB}${treatPathSep(change.newPath)}`
-      }
-      break
-    case 'delete':
-      if (keepSide(change.oldMode, change.oldPath, scopes)) {
-        yield `${DELETION}${TAB}${treatPathSep(change.oldPath)}`
-      }
-      break
-    case 'modify':
-      if (
-        change.oldMode !== GITLINK_MODE &&
-        keepSide(change.newMode, change.path, scopes)
-      ) {
-        yield `${MODIFICATION}${TAB}${treatPathSep(change.path)}`
-      }
-      break
-    case 'rename': {
-      const oldKept = keepSide(change.oldMode, change.oldPath, scopes)
-      const newKept = keepSide(change.newMode, change.newPath, scopes)
-      if (oldKept && newKept) {
-        yield `${RENAMED}${similarityPercent(change.similarity)}${TAB}${treatPathSep(change.oldPath)}${TAB}${treatPathSep(change.newPath)}`
-      } else if (newKept) {
-        yield `${ADDITION}${TAB}${treatPathSep(change.newPath)}`
-      } else if (oldKept) {
-        yield `${DELETION}${TAB}${treatPathSep(change.oldPath)}`
-      }
-      break
-    }
-  }
-}
-
-const GLOB_CHARS = /[*?[]/
-const REGEXP_SPECIALS = /[.+^${}()|\\\]]/g
-
-// Concrete repository paths are matched by directory prefix only — no
-// wildmatch, no leading-prefix normalisation. A path off the repository can
-// never carry a leading './' or '/' (treatPathSep and basePath already rule
-// those out), and treating `[` as a glob-class opener would misread real
-// path segments like an object folder named `Custom[1]__c`.
-const buildLiteralMatcher =
-  (specs: string[]) =>
-  (path: string): boolean =>
-    inScope(path, specs)
-
-// Git pathspec semantics: a literal pathspec matches by directory prefix; a
-// pathspec containing wildcards uses wildmatch where `*` also crosses `/`
-// (no `:(glob)` magic). Callers mix both shapes (e.g. flow translations use
-// `<source>/*.translation-meta.xml`).
-const buildPathspecMatcher = (specs: string[]): ((path: string) => boolean) => {
-  // Git normalizes leading `./` (and the `.//*` shape produced by the
-  // default `./` source dir) away before matching; repo paths never carry
-  // either prefix.
-  const normalized = specs.map(spec =>
-    spec.replace(/^(\.\/)+/, '').replace(/^\/+/, '')
-  )
-  const literals = normalized.filter(spec => !GLOB_CHARS.test(spec))
-  const globs = normalized
-    .filter(spec => GLOB_CHARS.test(spec))
-    .map(spec => {
-      const escaped = spec
-        .replace(REGEXP_SPECIALS, '\\$&')
-        .replace(/\*/g, '.*')
-        .replace(/\?/g, '.')
-      return new RegExp(`^${escaped}$`)
-    })
-  return (path: string): boolean =>
-    (literals.length > 0 && inScope(path, literals)) ||
-    globs.some(glob => glob.test(path))
 }
